Implement autosave functionality in conversation manager
